+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 Changes are grouped by the date they are merged to the main branch of the repository and are ordered from newest to oldest. Dates use the ISO 8601 extended calendar date format, i.e. YYYY-MM-DD.
 
+## 2024-10-25
+
+- Update test name in y-cruncher harness to not be the tuning used.
+- Update PugetBench harness to use the latest benchmark version available.
+
 ## 2024-10-18
 
 - Updated Cyberpunk 2077 harness to take screenshots of settings and results.